requirements.txt
# --- Core Data Science & ML ---
pandas
numpy
scikit-learn
imbalanced-learn
matplotlib
seaborn
altair<5

# --- Backend (FastAPI) ---
fastapi
uvicorn
pydantic

# --- Frontend (Streamlit) ---
streamlit>=1.31.0
requests

# --- Database & Environment ---
psycopg2-binary
python-dotenv
openpyxl
